--- src/components/tabs/tabs.jsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import Tab from './components/tab/tab';
-import {CITIES} from '../../const';
-import * as Type from '../../types';
-
-const Tabs = ({activeCityId}) => {
-  return (
-    <div className="tabs">
-      <section className="locations container">
-        <ul className="locations__list tabs__list">
-          {Object.values(CITIES).map((city) => (
-            <Tab key={city.id} city={city} activeCityId={activeCityId} />
-          ))}
-        </ul>
-      </section>
-    </div>
-  );
-};
-
-Tabs.propTypes = {
-  activeCityId: Type.id,
-};
-
-export default Tabs;
